backend/requirements.txt
# Autonomous AI Insurance Credentialing System - Python Dependencies

# LangChain and LangGraph
langchain>=0.1.0
langchain-openai>=0.0.2
langchain-core>=0.1.10
langgraph>=0.0.20
langsmith>=0.0.77

# OpenAI
openai>=1.6.1

# Telephony
twilio>=8.11.0
deepgram-sdk>=3.0.0
elevenlabs>=0.2.27

# Database
psycopg2-binary>=2.9.10
sqlalchemy>=2.0.25

# Web Framework
flask>=3.0.0
flask-cors>=4.0.0
flask-socketio>=5.3.5
python-socketio>=5.10.0
flask-jwt-extended>=4.6.0
bcrypt>=4.1.2

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0
aiohttp>=3.9.1
websockets>=12.0

# Data Processing
pydantic>=2.5.3
pydantic-settings>=2.1.0
tiktoken>=0.5.0

# Audio Processing
pydub>=0.25.1
numpy>=2.1.0

# Date/Time
python-dateutil>=2.8.2
pytz>=2023.3

# Logging and Monitoring
loguru>=0.7.2

# Environment
python-json-logger>=2.0.7

# Testing (optional)
pytest>=7.4.3
pytest-asyncio>=0.21.1
pytest-mock>=3.12.0
